color changes, Now using decimal type, webview works on ios, and some error handling on some of the http requests

+++ lib/components/addons_list_item.dart
@@ -1,9 +1,10 @@
+import 'package:decimal/decimal.dart';
 import 'package:delivery_app/models/food.dart';
 import 'package:flutter/material.dart';
 
 class AddonsListItem extends StatefulWidget {
   final String name;
-  final int price;
+  final Decimal price;
   Widget checkbox; 
 
   AddonsListItem({
